(SIMP-631) Updated to make sssd default in EL7
This change adds 'sss' as a option in /etc/nsswitch.conf as the default
for EL>6. If the SSSD daemon is not running, this will have no adverse
affects and it makes the system function properly out of the box for
systems that do have SSSD running.

Unfortunately, SSSD can have both local and remote accounts so the
situation is not as cut and dry as it was with the 'use_ldap' variable
in the past.

Luckily, nsswitch is quite well written and will not fail on items that
it does not recognize in the configuration file.

SIMP-631 #comment Update nsswitch to use sssd by default in EL>6

Change-Id: Ib438f0459c1d467617f3e75837e64f87179cbb95