(SIMP-1435) Fix `swappiness` to use `sysctl` type
Before this commit, the `simplib::swappiness` class still used the
`sysctl::value` define instead of the `sysctl` type from
`augeasproviders_systctl`.  This resulted in a SyntaxError that broke
the catalog when the `absolute_swappiness` was used.

This commit fixes the issue by convering the class and its tests from
using `sysctl::value` to `sysctl`.  It also clarifies the role of the
file `/usr/local/sbin/dynamic_swappiness.rb` by ensuring it is absent
from hosts that are configured to use an absolute `swappiness` value.

SIMP-1435 #close #comment Fix `swappiness` to use `sysctl` type

Change-Id: I93c7cf3bfffc925fccc6fb11cc7d1cabf04864b6