(SIMP-1694) Add the simplib::lookup function
This enhances the inbuilt lookup() function to check and see if the
passed value is defined in the global scope as a variable prior to
falling back to a normal lookup.

This means that any value that is set via an ENC or a standard class
declaration will be properly found.

SIMP-1694 #close