(SIMP-2937) Use PATH lookup for simp_version's rpm call
The `simp_version` fact executes `rpm(8)` to find the version for SIMP.
Previously this was done with a hard-coded path, `/bin/rpm`, but this
broke on systems where the executable was installed elsewhere.  This
changes the call to set up PATH in the subshell before executing the
command.

SIMP-2937 #close