(SIMP-2992) Added simplib::inspect debugger
The original inspect() function is no longer user-friendly enough for
debugging.

This function has been added to actually print the value during the
Puppet run as a notify resource.

Care was taken to try to make it easy to note that this is for
debugging.

SIMP-2992 #comment Needed during debugging of array sets