(SIMP-3115) Added wrapper stages to simplib
There have been several situations where various SIMP components really
needed to be wrapped in a Puppet stage to protect the rest of the
system from a particularly aggressive action (usually one that requires
a reboot).

This change adds a couple of SIMP stages that wrap the stdlib stages to
provide a buffer zone for SIMP-specific activities. This is not a
perfect solution but it will work better than just trying to tag onto
the existing stages that don't actually map to our use cases.

SIMP-3115 #comment Added buffer stages for module locking