(SIMP-3280) Add SHA256-based option to ip_to_cron/rand_cron
SIMP-3280 specifies to replace the IP-modulus-based algorithm in
rand_cron (and thus ip_to_cron) with a SHA256-based algorithm.
This request was specifically meant to address a deficiency with
the randomness of client puppet agent runs scheduled by
pupmod::agent::cron in certain circumstances. (By default,
pupmod::agent::cron uses rand_cron to determine the cron minute
for client's puppet agent run from the client's IP address.)

Since, empirical analsysis of both the original and proposed
algorithms for sets of servers in which both the number of IP
addresses is larger than the cron range and the IP addresses are
assigned linearly, the IP-modulus algorithm provides a better
distribution across minutes in an hour, this PR instead provides
both algorithms and allows the user to select the algorithm
appropriate for their site. A corresponding change will be made
to pupmod-simp-pupmod, to expose the new option.

Note that the crc32-based algorithm for non-IP input to rand_cron
will still be available, for backward compatibility.