(SIMP-4214) simplib: Create ipa fact
Created ipa fact that provides the IPA domain and server to which
a host is connected, when the host is joined to the IPA domain
AND is able to communicate with the IPA server.

As written, this fact cannot to distinguish between the following cases:

* No connectivity to the IPA server when joined to an IPA domain.
* The host is no longer joined to an IPA domain.

Addressing this deficiency is not trivial and will need to be done
in a follow-on ticket.

NOTE: At first blush, it seems like the fact could have been written
to pull information locally from the IPA configuration file,
/etc/ipa/default.conf.  However, the information in this file is not
definitive. Specifically, if the client has been removed from the IPA
domain using the IPA server GUI instead of running the IPA client
uninstall, this file will still exist but be inapplicable.

SIMP-4214 #close
SIMP-4215 #close
SIMP-4162 #comment code, unit, and acceptance test