(SIMP-4955) Fix boot_dir_uuid fact
There was an issue with the `boot_dir_uuid` fact where it would throw an
error if running on a system without a `/boot` partition.

This shows on `docker` tests but could also happen with headless nodes.

A docker-based test was added to the stack as a nodeset.

SIMP-4955 #close