(SIMP-6515) Defer to inbuilt fips_enable fact
Some version of facter have a built-in 'fips_enabled' fact. We now defer
to that one prior to adding our own fact.

Note: I was unable to determine how to properly unit test this but,
since the entire change is just checking for one item in Facter prior to
creating the fact, this isn't vastly different from the previous code
version (which also didn't have a unit test)

SIMP-6591 #close