Allow root to run `puppet lookup --compile` (#273)
#205 broke `puppet lookup --compile`, which is essential to troubleshoot Hiera lookups when using the
top-scope manifest variables in the stock manifests/site.pp. The defaults introduced in that patch
assumed that the Puppet user is the user executing the command/process.  If the user is root using the
default umask of a SIMP server, this result in files that the Puppet Server's puppet user will not be able 
to read.

This patch uses the Puppet settings to look up the configured Puppet user instead of the owner of the
current process.

Fixes #267